feat: add product persistence and APIs for create, get all, and get by id using mapper layer

+++ src/main/java/com/nivora/mart/gateway/IProductGateway.java
@@ -9,5 +9,5 @@ public interface IProductGateway {
 
     List<ProductDTO> getAllProducts() throws IOException;
 
-    ProductDTO getProductByID(int id) throws  IOException;
+    ProductDTO getProductByID(long id) throws  IOException;
 }
